Develop

+++ core/Group/Cron/Cron.php
@@ -4,6 +4,7 @@ namespace Group\Cron;
 
 use Group\Cron\ParseCrontab;
 use Group\App\App;
+use CronFileCache as FileCache;
 use swoole_process;
 use swoole_table;
 
@@ -31,23 +32,24 @@ class Cron
 
     protected $table;
 
-    protected $max_handle;
+    protected $timezone;
 
     protected $daemon = false;
 
     protected $help = "
-\033[34m
-   __   _ __   ___   __  __  _____   
- /'_ `\/\`'__\/ __`\/\ \/\ \/\ '__`\ 
-/\ \L\ \ \ \//\ \L\ \ \ \_\ \ \ \L\ \
-\ \____ \ \_\\ \____/\ \____/\ \ ,__/
- \/___L\ \/_/ \/___/  \/___/  \ \ \/ 
-   /\____/                     \ \_\ 
-   \_/__/                       \/_/ 
-
+\033[35m                                                                  
+                                   ____                            
+                                  /\  _`\                          
+   __   _ __   ___   __  __  _____\ \ \/\_\  _ __   ___     ___    
+ /'_ `\/\`'__\/ __`\/\ \/\ \/\ '__`\ \ \/_/_/\`'__\/ __`\ /' _ `\  
+/\ \L\ \ \ \//\ \L\ \ \ \_\ \ \ \L\ \ \ \L\ \ \ \//\ \L\ \/\ \/\ \ 
+\ \____ \ \_\\ \____/\ \____/\ \ ,__/\ \____/\ \_\\ \____/\ \_\ \_\
+ \/___L\ \/_/ \/___/  \/___/  \ \ \/  \/___/  \/_/ \/___/  \/_/\/_/
+   /\____/                     \ \_\                               
+   \_/__/                       \/_/                               
 \033[0m
 \033[31m 使用帮助: \033[0m
-\033[33m Usage: app/cron [start|restart|stop|status|exec (cron name)|rejob (cron name)]|server \033[0m
+\033[33m Usage: app/cron [start|restart|stop|status|exec (cron name)|server \033[0m
 ";
     /**
      * 初始化环境
@@ -55,16 +57,19 @@ class Cron
      */
     public function __construct($argv, $loader)
     {
-        $this->cacheDir = \Config::get('cron::cache_dir') ? : 'runtime/cron';
-        $this->tickTime = \Config::get('cron::tick_time') ? : 2;
+        $this->tickTime = 1;
         $this->argv = $argv;
         $this->loader = $loader;
         $this->jobs = \Config::get('cron::job');
         $this->workerNum = count($this->jobs);
-        $this->logDir = \Config::get("cron::log_dir");
+        $this->logDir = \Config::get("cron::log_dir") ? : 'runtime/cron';
+        $this->cacheDir = $this->logDir;
+        $this->timezone = \Config::get("cron::timezone") ? : 'PRC';
         $this->daemon = \Config::get("cron::daemon") ? : false;
-        $this->max_handle = \Config::get("cron::max_handle") ? : 10;
         \Log::$cacheDir = $this->logDir;
+        $this->logDir = __FILEROOT__.$this->logDir;
+
+        date_default_timezone_set($this->timezone);
     }
 
     /**
@@ -92,7 +97,8 @@ class Cron
         //设置信号
         $this->setSignal();
 
-        swoole_timer_tick($this->tickTime * 1000, function($timerId) {
+        $this->linstenId = swoole_timer_tick($this->tickTime * 1000, function($timerId) {
+            $ctime = time();
             foreach ($this->jobs as $key => $job) {
                 $worker = $this->table->get($job['name'].'_worker');
                 $worker = json_decode($worker[$job['name'].'_worker'], true);
@@ -103,11 +109,14 @@ class Cron
                     $this->newProcess($key);
                 }
 
+                $this->workers[$job['name']]['job']['ctime'] = time();
                 $this->workers[$job['name']]['process']->write(json_encode($this->workers[$job['name']]['job']));
             }
         });
 
         $this->setPid();
+
+        FileCache::set('linsten_id', $this->linstenId, $this->cacheDir);
     }
 
     public function status()
@@ -136,15 +145,21 @@ class Cron
 
         if (!empty($pid) && $pid) {
             if (swoole_process::kill($pid, 0)) {
+                //停止tick
+                //swoole_process::kill($pid, SIGUSR2);
+                FileCache::remove('linsten_id', $this->cacheDir);
+
+                $workIds = $this->findWorkIds($this->logDir);
                 //杀掉worker进程
-                foreach ($this->jobs as $job) {
-                    $work_id = \FileCache::get('work_id', $this->cacheDir."/".$job['name']);
+                foreach ($workIds as $work_id) {
                     if (is_array($work_id)) {
                         //向子进程发送退出命令,结束完当前任务后退出
                         try {
-                            swoole_process::kill($work_id[0], SIGTERM);
+                            if (swoole_process::kill($work_id[0], 0)) {
+                                swoole_process::kill($work_id[0], SIGTERM);
+                            }
                         } catch (Exception $e) {
-                            \Log::info("进程{$work_id}不存在", [], 'cron.stop');
+                            \Log::info("进程{$work_id[0]}不存在", [], 'cron.stop');
                         }
                     }
                 }
@@ -160,24 +175,46 @@ class Cron
     {
         //子进程结束时主进程收到的信号
         swoole_process::signal(SIGCHLD, function ($signo) {
+
             //kill掉所有worker进程 必须为false，非阻塞模式
             static $worker_count = 0;
             while($ret = swoole_process::wait(false)) {
                 $worker_count++;
                 $workerNum = $this->table->get('workers_num');
-                \Log::info("PID={$ret['pid']}worker进程退出!", [$workerNum, $signo], 'cron');
+                \Log::info("PID={$ret['pid']}worker进程退出!", [$workerNum, $signo, $worker_count], 'cron');
+
+                //看下是哪个进程退出了 清除worker信息
+                $job = $this->getJobByPid($ret['pid']);
+                if ($job && file_exists($this->logDir."/linsten_id")) {
+                    $this->table->set($job['name'].'_worker', [$job['name'].'_worker' => json_encode([])]);
+                    $this->removeWorkerPid($job['name']);
+
+                    $this->table->incr('workers_num', 'workers_num');
+                    $workerNum['workers_num']++;
+                }
 
                 if ($worker_count >= $workerNum['workers_num']){
-                    \Log::info("主进程退出!", [], 'cron');
+                    \Log::info("主进程退出!", [$workerNum], 'cron');
                     foreach ($this->jobs as $job) {
-                        @unlink($this->logDir."/".$job['name']."/work_id");
+                        FileCache::remove('work_id', $this->cacheDir."/".$job['name']);
+                        FileCache::remove('cronAdmin', $this->cacheDir."/".$job['name']);
                     }
                     
-                    @unlink($this->logDir."/pid");
+                    FileCache::remove('pid', $this->cacheDir);
                     swoole_process::kill($this->getPid(), SIGKILL); 
                 }
             }
         });
+
+        // swoole_process::signal(SIGUSR2, function ($signo) {
+        //     //主进程停止tick
+        //     \Log::info("主进程tick停止!", [], 'cron');
+        //     $linstenId = FileCache::get('linsten_id', $this->cacheDir);
+        //     if ($linstenId) {
+        //         swoole_timer_clear($linstenId);
+        //         FileCache::remove('linsten_id', $this->cacheDir);
+        //     }
+        // });
     }
 
     /**
@@ -241,25 +278,6 @@ class Cron
         exit("job不存在\n");
     }
 
-    public function rejob()
-    {
-        $argv = $this->argv;
-        $jobName = isset($argv[2]) ? $argv[2] :'';
-        foreach ($this->jobs as $job) {
-            if ($job['name'] == $jobName) {
-                $worker = \FileCache::get('cronAdmin', $this->cacheDir."/".$jobName);
-                if (isset($worker[0])) {
-                    $processPid = $worker[0]['pid'];
-                    exec("kill -USR1 {$processPid}");
-                    exit("{$jobName}重启完成\n");
-                }
-            }
-            continue;
-        }
-
-        exit("job不存在\n");
-    }
-
     public function workerCallBack(swoole_process $worker)
     {   
         $this->init();
@@ -268,27 +286,25 @@ class Cron
             $recv = $worker->read(); 
             $recv = json_decode($recv, true);
             if (!is_array($recv)) return;
-            
+
             $this->bindTick($recv);
         });
 
         //接受重启的信号
-        swoole_process::signal(SIGUSR1, function ($signo) use ($worker) {
-            $pid = $worker->pid;
-            foreach ($this->jobs as $job) {
-                $workers = \FileCache::get('cronAdmin', $this->cacheDir."/".$job['name']);
-                foreach ($workers as $worker) {
-                    if ($worker['pid'] == $pid) {
-                        $timerId = isset($worker['timerId']) ? $worker['timerId'] : 0;
-                        $this->restartJob($worker['job']);
-                    }
-                }
-            }
-        });
+        // swoole_process::signal(SIGUSR1, function ($signo) use ($worker) {
+        //     $pid = $worker->pid;
+        //     foreach ($this->jobs as $job) {
+        //         $workers = FileCache::get('work_id', $this->cacheDir."/".$job['name']);
+        //         if (isset($workers[0]) && $workers[0] == $pid) {
+        //             $this->restartJob(['name' => $job['name'], 'workId' => $pid]);
+        //         }
+        //     }
+        // });
 
         //接受退出的信号
         swoole_process::signal(SIGTERM, function ($signo) use ($worker) {
-            $worker->exit();
+            \Log::info("work 退出信号", [], 'cron');
+            $worker->exit(0);
         });
     }
 
@@ -300,22 +316,12 @@ class Cron
     {
         set_time_limit(0);
 
-        $count = 0;
-        while ($count <= $this->max_handle) {
-            $timer = ParseCrontab::parse($job['time']);
-            if (is_null($timer)) return;
+        $timer = ParseCrontab::parse($job['time'], $this->timezone, $job['ctime']);
+        if (is_null($timer)) return;
 
-            $job['timer'] = $timer;
-            $start = microtime(true);
-            $this->jobStart($job);
-            $end = microtime(true);
-            $timer -= intval($end - $start);
+        $job['timer'] = $timer;
 
-            if (sleep($timer) > 0) {
-               return; 
-            }
-            $count++;
-        }
+        $this->jobStart($job);
 
         $this->restartJob($job);
     }
@@ -337,7 +343,7 @@ class Cron
     private function setWorkerPid($pid, $jobName)
     {   
         $dir = $this->cacheDir."/".$jobName;
-        \FileCache::set('work_id', [$pid], $dir);
+        FileCache::set('work_id', [$pid], $dir);
     }
 
     /**
@@ -345,58 +351,61 @@ class Cron
      *
      * @param pid int
      */
-    private function removeWorkerPid($pid, $jobName)
+    private function removeWorkerPid($jobName)
     {   
-        $dir = $this->cacheDir."/".$jobName;
-        @unlink($this->cacheDir."/".$jobName."/work_id");
+        FileCache::remove('work_id', $this->cacheDir."/".$jobName);
     }
 
     public function setPid()
     {
         $pid = posix_getpid();
-        $parts = explode('/', $this->cacheDir."/pid");
-        $file = array_pop($parts);
-        $dir = '';
-        foreach ($parts as $part) {
-            if (!is_dir($dir .= "$part/")) {
-                 mkdir($dir);
-            }
-        }
-        file_put_contents("$dir/$file", $pid);
+        FileCache::set('pid', $pid, $this->cacheDir);
     }
 
     public function getPid()
+    {   
+        return FileCache::get('pid', $this->cacheDir);
+    }
+
+    private function getJobByPid($pid)
     {
-        if (file_exists($this->cacheDir."/pid"))
-        return file_get_contents($this->cacheDir."/pid");
+        //遍历目录
+        foreach ($this->jobs as $key => $one) {
+            if ($pid == $one['workId']) {
+                return $one;
+            }
+        }
+
+        return false;
     }
 
     private function jobStart($job)
     {
         $worker = $this->table->get($job['name'].'_worker');
         $worker = json_decode($worker[$job['name'].'_worker'], true);
-        $worker['startTime'] = date('Y-m-d H:i:s', time());
+        $worker['startTime'] = date('Y-m-d H:i:s', $job['ctime']);
         $worker['timer'] = intval($job['timer']);
-        $worker['nextTime'] = date('Y-m-d H:i:s', time() + intval($job['timer']));
+        $worker['nextTime'] = date('Y-m-d H:i:s', $job['ctime'] + intval($job['timer']));
         $this->table->set($job['name'].'_worker', [$job['name'].'_worker' => json_encode($worker)]);
 
-        \FileCache::set('cronAdmin', [$worker], $this->cacheDir."/".$job['name']);
+        FileCache::set('cronAdmin', [$worker], $this->cacheDir."/".$job['name']);
   
         \Log::info('定时任务启动'.$job['name'], [], 'cron.start');
 
         //先执行一次任务
-        call_user_func_array([new $job['command'], 'handle'], []);
+        try {
+            call_user_func_array([new $job['command'], 'handle'], []);
+        } catch (\Exception $e) {
+           \Log::error('定时任务执行失败'.$job['name'], [$e->getMessage()], 'cron.error'); 
+        }
     }
 
     private function restartJob($job)
     {
         foreach ($this->jobs as $key => $one) {
             if ($one['name'] == $job['name']) {
-                $this->table->set($job['name'].'_worker', [$job['name'].'_worker' => json_encode([])]);
-                $this->table->incr('workers_num', 'workers_num');
-
-                $this->removeWorkerPid($job['workId'], $job['name']);
-                swoole_process::kill($job['workId'], SIGKILL);
+                $this->removeWorkerPid($job['name']);
+                swoole_process::kill($job['workId'], SIGTERM);
                 break;
             }
         }
@@ -409,6 +418,8 @@ class Cron
 
         $this->setWorkerPid($processPid, $this->jobs[$i]['name']);
         
+        //$this->table->incr('workers_num', 'workers_num');
+
         $this->jobs[$i]['workId'] = $processPid;
         $this->workers[$this->jobs[$i]['name']] = [
             'process' => $process,
@@ -425,6 +436,33 @@ class Cron
         $this->table->set($this->jobs[$i]['name'].'_worker', [$this->jobs[$i]['name'].'_worker' => json_encode($worker)]);
 
         \Log::info("工作worker{$processPid}启动", [$this->jobs[$i]['name']], 'cron.work');
+    }
+
+    private  function findWorkIds($fileDir, $data = [])
+    {
+        if (is_dir($fileDir)) {
+            $dir = opendir($fileDir);
+            if (!$dir) {
+                return $data;
+            }
+            while (($file = readdir($dir)) !== false)
+            {   
+                $file = explode(".", $file);
+                $fileName = $file[0];
+
+                if ($fileName == "work_id") {
+                    $d = file_get_contents($fileDir."/".$fileName);
+                    if ($d) {
+                        $data[] = json_decode($d, true);
+                    }
+                } else if ($fileName) {
+                    $data = $this->findWorkIds($fileDir."/".$fileName, $data);
+                }
+            }
+            closedir($dir);
+        }
+
+        return $data;
     }
 
     private function init()
